requirements.txt
streamlit
python-dotenv
langchain
langchain-community
cohere
faiss-cpu
pypdf
sentence-transformers
tqdm
